main <- develop_yj

+++ src/main/java/com/movie/MovieReview/post/entitiy/Post.java
@@ -32,8 +32,9 @@ public class Post extends BaseTimeEntity {
 
     @Transient
     private boolean isLiked;
+
     @Transient
-    private Long likesCount;
+    private int likesCount;
 
     @Setter
     private int commentCnt;
